feat: scene change and go to main scene when click exit buton
