+++ Assets/Seoyoon/Scenes/LoadingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 62d211e427c3fbd4092f232104158c01
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
