+++ Assets/Seoyoon/Scenes/TitleProto.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b475850dbd763a42ac471194c6f302c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
