+++ Assets/Seoyoon/Scripts/Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 14970a9aeafb11b48b23e43c6d12f34c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
