+++ Assets/Seoyoon/Scripts/Manager/InGameUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6361eb737cb2ce46b12432bb4ac8176
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
