+++ Assets/Seoyoon/Scripts/Manager/LoadingManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25382038ffee0cf4891af57e143b4e83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
