+++ Assets/Seoyoon/prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f3b87460e983714596c44ed4fc6ab28
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
