--- Assets/YS/1_Script/Manager/SettingManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dcd792499309aa543a8de807d9945ce2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
