--- Assets/YS/1_Script/Manager/TitleManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbb55420418987c479b3cfb12bcab2e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
