+++ Assets/YS/Resources/BGFXShader.mat
@@ -58,7 +58,7 @@ Material:
     m_Floats:
     - _A: 0
     - _BumpScale: 1
-    - _CurTime: 1
+    - _CurTime: 0.99999934
     - _Cutoff: 0.5
     - _DetailNormalMapScale: 1
     - _DstBlend: 0
@@ -66,7 +66,7 @@ Material:
     - _GlossMapScale: 1
     - _Glossiness: 0.5
     - _GlossyReflections: 1
-    - _IsIn: 0
+    - _IsIn: 1
     - _Metallic: 0
     - _Mode: 0
     - _OcclusionStrength: 1
