+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Music
   - Dialog
   - Preview
+  - loadingText
   layers:
   - Default
   - TransparentFX
